# Improve the performance of skip_until_any