--- a/pest/Cargo.toml
+++ b/pest/Cargo.toml
@@ -10,9 +10,6 @@ categories = ["parsing"]
 license = "MIT/Apache-2.0"
 readme = "_README.md"
 
-[dependencies]
-memchr = "^2.0"
-
 [badges]
 codecov = { repository = "pest-parser/pest" }
 maintenance = { status = "actively-developed" }