@jstnlef 2018-03-29
Here are the new numbers when testing with the requested changes and rust 1.25
```
running 4 tests
test apache ... bench:     598,335 ns/iter (+/- 297,090) = 212 MB/s
test basic  ... bench:       1,377 ns/iter (+/- 710) = 55 MB/s
test canada ... bench:  29,317,882 ns/iter (+/- 3,546,155) = 76 MB/s
test data   ... bench:      48,559 ns/iter (+/- 24,376) = 190 MB/s
```