@jstnlef 2018-03-29
You mean specifically the one benchmark in `pest` right? Not the ones in `pest_grammar`