@jstnlef 2018-03-30
Deleted the benchmarks